Added three new functions.
Pull_data pulls all of the data available. To do this it uses get_ai_channels, which in turn uses get_site_types. This code was mostly taken from other files in acq400_hapi, so these functions  can be used to reduce the size of those files and stop code repetition.