Fixed unreliable "3 PV's vs one triplet" issue:  def get_stl(stl):      with open (stl, "r") as stl_file: @@ -35,10 +35,11 @@ def get_stl(stl):
 def set_gpg(args, uut):
     uut.s0.gpg_enable = 0
-    uut.s0.GPG_TRG = 'external'
-    uut.s0.GPG_TRG_DX = 'd0'
-    uut.s0.GPG_TRG_SENSE = 'rising'
+#    uut.s0.GPG_TRG = 'external'
+#    uut.s0.GPG_TRG_DX = 'd0'
+#    uut.s0.GPG_TRG_SENSE = 'rising'
     uut.s0.GPG_MODE = 'LOOPWAIT'
+    uut.s0.gpg_trg = '1,0,1'

Also, fix broken contract with acq400_upload
-import acq400_upload
+import acq400_fullshot

this is as a result of bd46fe2
git mv acq400_upload to acq400_fullshot (the kitchen sink)
create acq400_upload (just upload).

So this client was finding acq400_upload, but this now doesn't have the
features it needed. gpg_load_oneshot.py is the ONLY client of
acq400_upload.py

 On branch master
	modified:   user_apps/acq400/gpg_load_oneshot.py